+++ src/imageLoader/colorSpaceConverters/convertYBRFullByPlane.js
@@ -1,4 +1,4 @@
-export default function (imageFrame, rgbaBuffer) {
+export default function(imageFrame, rgbaBuffer) {
   if (imageFrame === undefined) {
     throw new Error('decodeRGB: ybrBuffer must not be undefined');
   }
@@ -7,9 +7,13 @@ export default function (imageFrame, rgbaBuffer) {
   }
 
   const numPixels = imageFrame.length / 3;
+
   let rgbaIndex = 0;
+
   let yIndex = 0;
+
   let cbIndex = numPixels;
+
   let crIndex = numPixels * 2;
 
   for (let i = 0; i < numPixels; i++) {
@@ -17,9 +21,9 @@ export default function (imageFrame, rgbaBuffer) {
     const cb = imageFrame[cbIndex++];
     const cr = imageFrame[crIndex++];
 
-    rgbaBuffer[rgbaIndex++] = y + 1.40200 * (cr - 128);// red
+    rgbaBuffer[rgbaIndex++] = y + 1.402 * (cr - 128); // red
     rgbaBuffer[rgbaIndex++] = y - 0.34414 * (cb - 128) - 0.71414 * (cr - 128); // green
-    rgbaBuffer[rgbaIndex++] = y + 1.77200 * (cb - 128); // blue
+    rgbaBuffer[rgbaIndex++] = y + 1.772 * (cb - 128); // blue
     rgbaBuffer[rgbaIndex++] = 255; // alpha
   }
 }