+++ src/imageLoader/createImage.js
@@ -8,14 +8,16 @@ import isJPEGBaseline8BitColor from './isJPEGBaseline8BitColor.js';
 
 let lastImageIdDrawn = '';
 
-function isModalityLUTForDisplay (sopClassUid) {
+function isModalityLUTForDisplay(sopClassUid) {
   // special case for XA and XRF
   // https://groups.google.com/forum/#!searchin/comp.protocols.dicom/Modality$20LUT$20XA/comp.protocols.dicom/UBxhOZ2anJ0/D0R_QP8V2wIJ
-  return sopClassUid !== '1.2.840.10008.5.1.4.1.1.12.1' && // XA
-         sopClassUid !== '1.2.840.10008.5.1.4.1.1.12.2.1'; // XRF
+  return (
+    sopClassUid !== '1.2.840.10008.5.1.4.1.1.12.1' && // XA
+    sopClassUid !== '1.2.840.10008.5.1.4.1.1.12.2.1'
+  ); // XRF
 }
 
-function convertToIntPixelData (floatPixelData) {
+function convertToIntPixelData(floatPixelData) {
   const floatMinMax = getMinMax(floatPixelData);
   const floatRange = Math.abs(floatMinMax.max - floatMinMax.min);
   const intRange = 65535;
@@ -23,7 +25,9 @@ function convertToIntPixelData (floatPixelData) {
   const intercept = floatMinMax.min;
   const numPixels = floatPixelData.length;
   const intPixelData = new Uint16Array(numPixels);
+
   let min = 65535;
+
   let max = 0;
 
   for (let i = 0; i < numPixels; i++) {
@@ -39,7 +43,7 @@ function convertToIntPixelData (floatPixelData) {
     max,
     intPixelData,
     slope,
-    intercept
+    intercept,
   };
 }
 
@@ -48,7 +52,7 @@ function convertToIntPixelData (floatPixelData) {
  * can transfer array buffers but not typed arrays
  * @param imageFrame
  */
-function setPixelDataType (imageFrame) {
+function setPixelDataType(imageFrame) {
   if (imageFrame.bitsAllocated === 32) {
     imageFrame.pixelData = new Float32Array(imageFrame.pixelData);
   } else if (imageFrame.bitsAllocated === 16) {
@@ -62,8 +66,7 @@ function setPixelDataType (imageFrame) {
   }
 }
 
-function createImage (imageId, pixelData, transferSyntax, options) {
-
+function createImage(imageId, pixelData, transferSyntax, options = {}) {
   if (!pixelData || !pixelData.length) {
     return Promise.reject(new Error('The file does not contain image data.'));
   }
@@ -71,14 +74,59 @@ function createImage (imageId, pixelData, transferSyntax, options) {
   const { cornerstone } = external;
   const canvas = document.createElement('canvas');
   const imageFrame = getImageFrame(imageId);
-  const decodePromise = decodeImageFrame(imageFrame, transferSyntax, pixelData, canvas, options);
+
+  const decodePromise = decodeImageFrame(
+    imageFrame,
+    transferSyntax,
+    pixelData,
+    canvas,
+    options
+  );
 
   return new Promise((resolve, reject) => {
-    decodePromise.then(function (imageFrame) {
-      const imagePlaneModule = cornerstone.metaData.get('imagePlaneModule', imageId) || {};
-      const voiLutModule = cornerstone.metaData.get('voiLutModule', imageId) || {};
-      const modalityLutModule = cornerstone.metaData.get('modalityLutModule', imageId) || {};
-      const sopCommonModule = cornerstone.metaData.get('sopCommonModule', imageId) || {};
+    // eslint-disable-next-line complexity
+    decodePromise.then(function handleDecodeResponse(imageFrame) {
+      // If we have a target buffer that was written to in the
+      // Decode task, point the image to it here.
+      // We can't have done it within the thread incase it was a SharedArrayBuffer.
+      if (options.targetBuffer) {
+        const { arrayBuffer, offset, length, type } = options.targetBuffer;
+
+        let TypedArrayConstructor;
+
+        switch (type) {
+          case 'Uint8Array':
+            TypedArrayConstructor = Uint8Array;
+            break;
+          case 'Uint16Array':
+            TypedArrayConstructor = Uint16Array;
+            break;
+          case 'Float32Array':
+            TypedArrayConstructor = Float32Array;
+            break;
+          default:
+            throw new Error(
+              'target array for image does not have a valid type.'
+            );
+        }
+
+        const targetArray = new TypedArrayConstructor(
+          arrayBuffer,
+          offset,
+          length
+        );
+
+        imageFrame.pixelData = targetArray;
+      }
+
+      const imagePlaneModule =
+        cornerstone.metaData.get('imagePlaneModule', imageId) || {};
+      const voiLutModule =
+        cornerstone.metaData.get('voiLutModule', imageId) || {};
+      const modalityLutModule =
+        cornerstone.metaData.get('modalityLutModule', imageId) || {};
+      const sopCommonModule =
+        cornerstone.metaData.get('sopCommonModule', imageId) || {};
       const isColorImage = isColorImageFn(imageFrame.photometricInterpretation);
 
       // JPEGBaseline (8 bits) is already returning the pixel data in the right format (rgba)
@@ -93,7 +141,10 @@ function createImage (imageId, pixelData, transferSyntax, options) {
           canvas.width = imageFrame.columns;
 
           const context = canvas.getContext('2d');
-          const imageData = context.createImageData(imageFrame.columns, imageFrame.rows);
+          const imageData = context.createImageData(
+            imageFrame.columns,
+            imageFrame.rows
+          );
 
           convertColorSpace(imageFrame, imageData);
           imageFrame.imageData = imageData;
@@ -110,22 +161,30 @@ function createImage (imageId, pixelData, transferSyntax, options) {
       const image = {
         imageId,
         color: isColorImage,
-        columnPixelSpacing: imagePlaneModule.pixelSpacing ? imagePlaneModule.pixelSpacing[1] : undefined,
+        columnPixelSpacing: imagePlaneModule.columnPixelSpacing,
         columns: imageFrame.columns,
         height: imageFrame.rows,
-        intercept: modalityLutModule.rescaleIntercept ? modalityLutModule.rescaleIntercept : 0,
+        intercept: modalityLutModule.rescaleIntercept
+          ? modalityLutModule.rescaleIntercept
+          : 0,
         invert: imageFrame.photometricInterpretation === 'MONOCHROME1',
         minPixelValue: imageFrame.smallestPixelValue,
         maxPixelValue: imageFrame.largestPixelValue,
-        rowPixelSpacing: imagePlaneModule.pixelSpacing ? imagePlaneModule.pixelSpacing[0] : undefined,
+        rowPixelSpacing: imagePlaneModule.rowPixelSpacing,
         rows: imageFrame.rows,
-        sizeInBytes: imageFrame.pixelData.length,
-        slope: modalityLutModule.rescaleSlope ? modalityLutModule.rescaleSlope : 1,
+        sizeInBytes: imageFrame.pixelData.byteLength,
+        slope: modalityLutModule.rescaleSlope
+          ? modalityLutModule.rescaleSlope
+          : 1,
         width: imageFrame.columns,
-        windowCenter: voiLutModule.windowCenter ? voiLutModule.windowCenter[0] : undefined,
-        windowWidth: voiLutModule.windowWidth ? voiLutModule.windowWidth[0] : undefined,
+        windowCenter: voiLutModule.windowCenter
+          ? voiLutModule.windowCenter[0]
+          : undefined,
+        windowWidth: voiLutModule.windowWidth
+          ? voiLutModule.windowWidth[0]
+          : undefined,
         decodeTimeInMS: imageFrame.decodeTimeInMS,
-        floatPixelData: undefined
+        floatPixelData: undefined,
       };
 
       // add function to return pixel data
@@ -144,7 +203,7 @@ function createImage (imageId, pixelData, transferSyntax, options) {
       }
 
       if (image.color) {
-        image.getCanvas = function () {
+        image.getCanvas = function() {
           if (lastImageIdDrawn === imageId) {
             return canvas;
           }
@@ -161,15 +220,19 @@ function createImage (imageId, pixelData, transferSyntax, options) {
       }
 
       // Modality LUT
-      if (modalityLutModule.modalityLUTSequence &&
+      if (
+        modalityLutModule.modalityLUTSequence &&
         modalityLutModule.modalityLUTSequence.length > 0 &&
-        isModalityLUTForDisplay(sopCommonModule.sopClassUID)) {
+        isModalityLUTForDisplay(sopCommonModule.sopClassUID)
+      ) {
         image.modalityLUT = modalityLutModule.modalityLUTSequence[0];
       }
 
       // VOI LUT
-      if (voiLutModule.voiLUTSequence &&
-        voiLutModule.voiLUTSequence.length > 0) {
+      if (
+        voiLutModule.voiLUTSequence &&
+        voiLutModule.voiLUTSequence.length > 0
+      ) {
         image.voiLUT = voiLutModule.voiLUTSequence[0];
       }