+++ src/imageLoader/wadouri/metaData/metaDataProvider.js
@@ -3,10 +3,11 @@ import getNumberValues from './getNumberValues.js';
 import parseImageId from '../parseImageId.js';
 import dataSetCacheManager from '../dataSetCacheManager.js';
 import getImagePixelModule from './getImagePixelModule.js';
+import getOverlayPlaneModule from './getOverlayPlaneModule.js';
 import getLUTs from './getLUTs.js';
 import getModalityLUTOutputPixelRepresentation from './getModalityLUTOutputPixelRepresentation.js';
 
-function metaDataProvider (type, imageId) {
+function metaDataProvider(type, imageId) {
   const { dicomParser } = external;
   const parsedImageId = parseImageId(imageId);
 
@@ -23,7 +24,7 @@ function metaDataProvider (type, imageId) {
       seriesNumber: dataSet.intString('x00200011'),
       studyInstanceUID: dataSet.string('x0020000d'),
       seriesDate: dicomParser.parseDA(dataSet.string('x00080021')),
-      seriesTime: dicomParser.parseTM(dataSet.string('x00080031') || '')
+      seriesTime: dicomParser.parseTM(dataSet.string('x00080031') || ''),
     };
   }
 
@@ -31,18 +32,18 @@ function metaDataProvider (type, imageId) {
     return {
       patientAge: dataSet.intString('x00101010'),
       patientSize: dataSet.floatString('x00101020'),
-      patientWeight: dataSet.floatString('x00101030')
+      patientWeight: dataSet.floatString('x00101030'),
     };
   }
 
   if (type === 'imagePlaneModule') {
-
     const imageOrientationPatient = getNumberValues(dataSet, 'x00200037', 6);
     const imagePositionPatient = getNumberValues(dataSet, 'x00200032', 3);
     const pixelSpacing = getNumberValues(dataSet, 'x00280030', 2);
 
-    let columnPixelSpacing = 1.0;
-    let rowPixelSpacing = 1.0;
+    let columnPixelSpacing = null;
+
+    let rowPixelSpacing = null;
 
     if (pixelSpacing) {
       rowPixelSpacing = pixelSpacing[0];
@@ -50,11 +51,20 @@ function metaDataProvider (type, imageId) {
     }
 
     let rowCosines = null;
+
     let columnCosines = null;
 
     if (imageOrientationPatient) {
-      rowCosines = [parseFloat(imageOrientationPatient[0]), parseFloat(imageOrientationPatient[1]), parseFloat(imageOrientationPatient[2])];
-      columnCosines = [parseFloat(imageOrientationPatient[3]), parseFloat(imageOrientationPatient[4]), parseFloat(imageOrientationPatient[5])];
+      rowCosines = [
+        parseFloat(imageOrientationPatient[0]),
+        parseFloat(imageOrientationPatient[1]),
+        parseFloat(imageOrientationPatient[2]),
+      ];
+      columnCosines = [
+        parseFloat(imageOrientationPatient[3]),
+        parseFloat(imageOrientationPatient[4]),
+        parseFloat(imageOrientationPatient[5]),
+      ];
     }
 
     return {
@@ -69,7 +79,7 @@ function metaDataProvider (type, imageId) {
       sliceLocation: dataSet.floatString('x00201041'),
       pixelSpacing,
       rowPixelSpacing,
-      columnPixelSpacing
+      columnPixelSpacing,
     };
   }
 
@@ -82,25 +92,32 @@ function metaDataProvider (type, imageId) {
       rescaleIntercept: dataSet.floatString('x00281052'),
       rescaleSlope: dataSet.floatString('x00281053'),
       rescaleType: dataSet.string('x00281054'),
-      modalityLUTSequence: getLUTs(dataSet.uint16('x00280103'), dataSet.elements.x00283000)
+      modalityLUTSequence: getLUTs(
+        dataSet.uint16('x00280103'),
+        dataSet.elements.x00283000
+      ),
     };
   }
 
   if (type === 'voiLutModule') {
-    const modalityLUTOutputPixelRepresentation = getModalityLUTOutputPixelRepresentation(dataSet);
-
+    const modalityLUTOutputPixelRepresentation = getModalityLUTOutputPixelRepresentation(
+      dataSet
+    );
 
     return {
       windowCenter: getNumberValues(dataSet, 'x00281050', 1),
       windowWidth: getNumberValues(dataSet, 'x00281051', 1),
-      voiLUTSequence: getLUTs(modalityLUTOutputPixelRepresentation, dataSet.elements.x00283010)
+      voiLUTSequence: getLUTs(
+        modalityLUTOutputPixelRepresentation,
+        dataSet.elements.x00283010
+      ),
     };
   }
 
   if (type === 'sopCommonModule') {
     return {
       sopClassUID: dataSet.string('x00080016'),
-      sopInstanceUID: dataSet.string('x00080018')
+      sopInstanceUID: dataSet.string('x00080018'),
     };
   }
 
@@ -111,16 +128,26 @@ function metaDataProvider (type, imageId) {
       return;
     }
 
-    const firstRadiopharmaceuticalInfoDataSet = radiopharmaceuticalInfo.items[0].dataSet;
-
+    const firstRadiopharmaceuticalInfoDataSet =
+      radiopharmaceuticalInfo.items[0].dataSet;
 
     return {
       radiopharmaceuticalInfo: {
-        radiopharmaceuticalStartTime: dicomParser.parseTM(firstRadiopharmaceuticalInfoDataSet.string('x00181072') || ''),
-        radionuclideTotalDose: firstRadiopharmaceuticalInfoDataSet.floatString('x00181074'),
-        radionuclideHalfLife: firstRadiopharmaceuticalInfoDataSet.floatString('x00181075')
-      }
+        radiopharmaceuticalStartTime: dicomParser.parseTM(
+          firstRadiopharmaceuticalInfoDataSet.string('x00181072') || ''
+        ),
+        radionuclideTotalDose: firstRadiopharmaceuticalInfoDataSet.floatString(
+          'x00181074'
+        ),
+        radionuclideHalfLife: firstRadiopharmaceuticalInfoDataSet.floatString(
+          'x00181075'
+        ),
+      },
     };
+  }
+
+  if (type === 'overlayPlaneModule') {
+    return getOverlayPlaneModule(dataSet);
   }
 }