+++ src/imageLoader/wadouri/parseImageId.js
@@ -1,8 +1,10 @@
-function parseImageId (imageId) {
+function parseImageId(imageId) {
   // build a url by parsing out the url scheme and frame index from the imageId
   const firstColonIndex = imageId.indexOf(':');
+
   let url = imageId.substring(firstColonIndex + 1);
   const frameIndex = url.indexOf('frame=');
+
   let frame;
 
   if (frameIndex !== -1) {
@@ -15,7 +17,7 @@ function parseImageId (imageId) {
   return {
     scheme: imageId.substr(0, firstColonIndex),
     url,
-    frame
+    frame,
   };
 }